refactor: loginUser
